# Backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Unora mobile application",
        "title": "Unora Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/Unora-app/Backend",
            "email": "support@unora.app"
        },
        "license": {
            "name": "Private",
            "url": "https://github.com/Unora-app/Backend"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/connections": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all active connections for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "connections"
                ],
                "summary": "Get connections",
                "responses": {
                    "200": {
                        "description": "Connections",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_UnoraApp_be_internal_discovery_dto.ConnectionResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/connections/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed information about a specific connection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "connections"
                ],
                "summary": "Get connection details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Connection ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Connection details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_discovery_dto.ConnectionDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Connection not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Terminate an active connection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "connections"
                ],
                "summary": "Terminate connection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Connection ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Connection terminated",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/discover/refresh-status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Check if discovery refresh is available",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discovery"
                ],
                "summary": "Get refresh status",
                "responses": {
                    "200": {
                        "description": "Refresh status",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_discovery_dto.RefreshStatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/interests": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Express interest in a user from a discovery card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interests"
                ],
                "summary": "Express interest",
                "parameters": [
                    {
                        "description": "Interest data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_discovery_dto.ExpressInterestRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Interest created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_discovery_dto.InterestResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request or already interested",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/interests/received": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get pending interests received by the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interests"
                ],
                "summary": "Get received interests",
                "responses": {
                    "200": {
                        "description": "Received interests",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_UnoraApp_be_internal_discovery_dto.InterestResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/interests/sent": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get interests sent by the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interests"
                ],
                "summary": "Get sent interests",
                "responses": {
                    "200": {
                        "description": "Sent interests",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_UnoraApp_be_internal_discovery_dto.InterestResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/servers/{serverType}/discover": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get current batch of discovery cards for a server type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discovery"
                ],
                "summary": "Get discovery cards",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server type (partner/friend/growth)",
                        "name": "serverType",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Discovery batch",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_discovery_dto.DiscoveryBatchResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/servers/{serverType}/discover/refresh": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Trigger a new discovery batch (respects global 24h cooldown)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discovery"
                ],
                "summary": "Refresh discovery",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server type (partner/friend/growth)",
                        "name": "serverType",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New batch",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_discovery_dto.DiscoveryBatchResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Refresh not available yet",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_UnoraApp_be_internal_discovery_dto.ConnectionDetailResponse": {
            "description": "Full connection details with streak and reveals",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "partner": {
                    "$ref": "#/definitions/github_com_UnoraApp_be_internal_discovery_dto.ConnectionPartner"
                },
                "reveals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_UnoraApp_be_internal_discovery_dto.RevealSummary"
                    }
                },
                "serverType": {
                    "type": "string",
                    "example": "partner"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "streak": {
                    "$ref": "#/definitions/github_com_UnoraApp_be_internal_discovery_dto.StreakSummary"
                },
                "streakDetails": {
                    "$ref": "#/definitions/github_com_UnoraApp_be_internal_discovery_dto.StreakDetailResponse"
                }
            }
        },
...
        "github_com_UnoraApp_be_internal_discovery_dto.ConnectionResponse": {
            "description": "Active connection between users",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "partner": {
                    "$ref": "#/definitions/github_com_UnoraApp_be_internal_discovery_dto.ConnectionPartner"
                },
                "serverType": {
                    "type": "string",
                    "example": "partner"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "streak": {
                    "$ref": "#/definitions/github_com_UnoraApp_be_internal_discovery_dto.StreakSummary"
                }
            }
        },
...
        "github_com_UnoraApp_be_internal_discovery_dto.ExpressInterestRequest": {
            "description": "Express interest in a discovery card",
            "type": "object",
            "required": [
                "discoveryCardId"
            ],
            "properties": {
                "discoveryCardId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "github_com_UnoraApp_be_internal_discovery_dto.RefreshStatusResponse": {
            "description": "Status of discovery refresh availability",
            "type": "object",
            "properties": {
                "canRefresh": {
                    "type": "boolean",
                    "example": true
                },
                "refreshAvailableAt": {
                    "type": "string",
                    "example": "2024-01-01T12:00:00Z"
                },
                "secondsUntilRefresh": {
                    "type": "integer",
                    "example": 3600
                }
            }
        },
...
        "github_com_UnoraApp_be_pkg_response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIError"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}